CHEF-4086: Added base_resource_url and s3_external_url configuration
By default base_resource_url is the atom 'host_header'. In this setting,
URL generation returned to the requestor will be based upon the value for
the Host header in the request.

You can set a string value for base_resource_url to override the URL
being sent back to the requestor. For example, "https://api.example.com"

Likewise, s3_external_url controls the presigned URLs for cookbook
artifacts. By default, it is set to the atom 'host_header' and will use
the Host header from the request to generate the presigned url.

This can also take a string value.

When configuring for S3 instead of bookshelf, this value should be set
to the same value found in s3_url.