Change default rabbitmq port 5672 => 8672
Although 5672 is the IANA registered port for rabbitmq, on Redhat-6
based systems the qpidd service runs on that port. To avoid conflicts
with this service as well as to avoid conflicts with system installed
rabbitmq, the default port for chef-server's rabbitmq is now
8672. This value was chosen by incrementing the first digit of the
default rabbitmq port until a non-IANA assigned port was found.

The two chef-server services that connect to rabbitmq are
chef-expander and erchef. Both render the rabbitmq port in their
respective config files which have restart notifications for the
service.

rabbitmqctl communicates with the rabbitmq server using Erlang
distribution and does not rely on the listen port of the rabbitmq
server.

[CHEF-4346]