Check for postgres running as part of guards in postgresql recipe
Add PgHelper to organize guards used in the postgresql recipe. With
these changes we skip resources that create dbs and db users if the db
isn't running. If the db _is_ running, then we run checks to make the
resources idempotent.

This change allows the following command sequence to succeed:

    chef-server-ctl stop
    chef-server-ctl reconfigure
    chef-server-ctl start

This is an important pattern for upgrades and for major config
changes, such as changing the port on which a service runs.