remove obsoletes reference.
this is bad copypasta from the chef project.  the chef project
needs to replace "chef-full" which used to be the name of the
artifact that was produced a long time ago and we changed the
name.  the chef-server artifact has always been named that, so it
never needed this line at all.

this fixes errors in yum like:

"Cannot install package chef-server-11.0.12-1.el6.x86_64. It is
obsoleted by installed package chef-server-11.0.10-1.el6.x86_64"

this cannot fix existing broken installations that have obsoletes
so users should probably use this for now (which uses rpm, not
yum to do the install):

curl -L https://www.opscode.com/chef/install.sh | bash -s -- -P server