Add Docker and SystemD support to runit.
While we are at it, stop detecting the init system by distro type and
version number, and check it heuristically by the filesystem structures
that are unique to each init system, while assuming that Docker
containers won't have a working init no matter what we would otherwise
detect.

Extending this to other init schemes and container types is left as an
exercise for the annoyed sysadmin.