Add a note about upgrading pre-11.0.8 installs and runit change
In 11.0.8 the name of the Chef server runit was changed from having
opscode in the name to using chef-server. It has been observed that
in some installs this change can leave behind orphaned processes.
This updated includes directions on how to avoid this.

Ideally we would add a prerm script to run these steps, but that hasn't
yet been done.