initctl stop chef-server-runsvdir leaves orphaned processes
The right thing to do would be to have 'kill signal HUP' in
the Upstart config so everything gets taken care of cleanly
when you use upstart to stop the runsvdir process. But that
stanza isn't available in Ubuntu 10.04. Without that stanza
Upstart sends a TERM which kills runsvdir and orphans all the
rest of its process tree.

This fix does the same thing Enterprise Chef server does. It
stops Chef services in a pre-stop script.