Vendor `chef-gem` and `libffi` software definitions
Ohai 7.2.0 was just released and has a dependency on the `ffi` gem which
ships with a native extension which requires `libffi`. The fix is to add
a proper dependency on `libffi` in the `chef-gem` software definition.

We can’t simply update to the latest ominibus-software as this project 
is still on Omnibus 2.0.x. The alternate is to vendor the affected 
software definitions tweaked for the Omnibus 2.0 DSL.