Update sqitch verify command used during upgrade
We try to run sqitch verify to determine if sqitch is on a box during
upgrade. With the recent lock down of postgres, this needs to be run as
the db_user or else it will fail with an error, making the upgrade
process think sqitch isn't on the box, when it is.

Added a comment about this issue in the code as well to clarify this
issue.

Also removed 2 from the possible return values, as verify won't return a
2 (that's an issue with the sqitch deploy command in some cases).